+++ pkg/query/query_builder.go
@@ -2,9 +2,6 @@ package query
 
 import (
 	"context"
-	"fmt"
-	"hash/fnv"
-	"io"
 	"sort"
 	"strconv"
 	"strings"
@@ -28,7 +25,7 @@ const sourceAUR = "aur"
 type Builder interface {
 	Len() int
 	Execute(ctx context.Context, dbExecutor db.Executor, pkgS []string)
-	Results(w io.Writer, dbExecutor db.Executor, verboseSearch SearchVerbosity) error
+	Results(dbExecutor db.Executor, verboseSearch SearchVerbosity) error
 	GetTargets(include, exclude intrange.IntRanges, otherExclude stringset.StringSet) ([]string, error)
 }
 
@@ -84,6 +81,7 @@ type abstractResults struct {
 	bottomUp        bool
 	metric          strutil.StringMetric
 	separateSources bool
+	sortBy          string
 
 	distanceCache       map[string]float64
 	separateSourceCache map[string]float64
@@ -91,80 +89,6 @@ type abstractResults struct {
 
 func (a *abstractResults) Len() int      { return len(a.results) }
 func (a *abstractResults) Swap(i, j int) { a.results[i], a.results[j] = a.results[j], a.results[i] }
-
-func (a *abstractResults) GetMetric(pkg *abstractResult) float64 {
-	if v, ok := a.distanceCache[pkg.name]; ok {
-		return v
-	}
-
-	if strings.EqualFold(pkg.name, a.search) {
-		return 1.0
-	}
-
-	sim := strutil.Similarity(pkg.name, a.search, a.metric)
-
-	for _, prov := range pkg.provides {
-		// If the package provides search, it's a perfect match
-		// AUR packages don't populate provides
-		candidate := strutil.Similarity(prov, a.search, a.metric) * 0.80
-		if candidate > sim {
-			sim = candidate
-		}
-	}
-
-	simDesc := strutil.Similarity(pkg.description, a.search, a.metric)
-
-	// slightly overweight sync sources by always giving them max popularity
-	popularity := 1.0
-	if pkg.source == sourceAUR {
-		popularity = 1 - (30 / (30 + float64(pkg.votes)))
-	}
-
-	sim = sim*0.5 + simDesc*0.2 + popularity*0.3
-
-	a.distanceCache[pkg.name] = sim
-
-	return sim
-}
-
-func (a *abstractResults) separateSourceScore(source string, score float64) float64 {
-	if !a.separateSources {
-		return 0
-	}
-
-	if score == 1.0 {
-		return 50
-	}
-
-	switch source {
-	case sourceAUR:
-		return 0
-	case "core":
-		return 40
-	case "extra":
-		return 30
-	case "community":
-		return 20
-	case "multilib":
-		return 10
-	}
-
-	if v, ok := a.separateSourceCache[source]; ok {
-		return v
-	}
-
-	h := fnv.New32a()
-	h.Write([]byte(source))
-	sourceScore := float64(int(h.Sum32())%9 + 2)
-	a.separateSourceCache[source] = sourceScore
-
-	return sourceScore
-}
-
-func (a *abstractResults) calculateMetric(pkg *abstractResult) float64 {
-	score := a.GetMetric(pkg)
-	return a.separateSourceScore(pkg.source, score) + score
-}
 
 func (a *abstractResults) Less(i, j int) bool {
 	pkgA := a.results[i]
@@ -192,10 +116,11 @@ func (s *SourceQueryBuilder) Execute(ctx context.Context, dbExecutor db.Executor
 	sortableResults := &abstractResults{
 		results:             []abstractResult{},
 		search:              strings.Join(pkgS, ""),
-		distanceCache:       map[string]float64{},
 		bottomUp:            s.bottomUp,
 		metric:              metric,
 		separateSources:     s.separateSources,
+		sortBy:              s.sortBy,
+		distanceCache:       map[string]float64{},
 		separateSourceCache: map[string]float64{},
 	}
 
@@ -266,10 +191,10 @@ func (s *SourceQueryBuilder) Execute(ctx context.Context, dbExecutor db.Executor
 	}
 }
 
-func (s *SourceQueryBuilder) Results(w io.Writer, dbExecutor db.Executor, verboseSearch SearchVerbosity) error {
+func (s *SourceQueryBuilder) Results(dbExecutor db.Executor, verboseSearch SearchVerbosity) error {
 	for i := range s.results {
 		if verboseSearch == Minimal {
-			_, _ = fmt.Fprintln(w, s.results[i].name)
+			s.logger.Println(s.results[i].name)
 			continue
 		}
 
@@ -292,7 +217,7 @@ func (s *SourceQueryBuilder) Results(w io.Writer, dbExecutor db.Executor, verbos
 			toPrint += syncPkgSearchString(syncPkg, dbExecutor, s.singleLineResults)
 		}
 
-		fmt.Fprintln(w, toPrint)
+		s.logger.Println(toPrint)
 	}
 
 	return nil
